backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
email-validator==2.1.0.post1
fastapi==0.115.12
firebase-admin==6.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
python-jose[cryptography]==3.3.0
openai==1.79.0
passlib==1.7.4
pydantic==2.11.4
pydantic_core==2.33.2
pydantic[email]
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
